+++ packages/network-controller/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add MegaETH Testnet V2 support ([#7272](https://github.com/MetaMask/core/pull/7272))
+  - Add `megaeth-testnet-v2` in method `getDefaultCustomNetworkConfigurationsByChainId`
+  - Add `megaeth-testnet-v2` in type `AdditionalDefaultNetwork`
+
 ### Changed
 
 - Update the default set of Infura networks to include Monad Testnet ([#7067](https://github.com/MetaMask/core/pull/7067))